seperate nested index tests
